+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Ability to set custom keybindings in config (#135)
+
+
 ## [1.7.2] - 2020-12-21
 ### Fixed
 - Viewing subscriptions after subscribing to a certain user page won't crash Amfora (#157)